feat: velocity improvement - chance option on modify - filter option on modify - cancel complete option on watchdog - blink until ground on watchdog - packets option on exempt grim 117 - refactored code
