+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/combat/velocity/mode/VelocityDexland.kt
@@ -1,0 +1,49 @@
+/*
+ * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
+ *
+ * Copyright (c) 2015 - 2024 CCBlueX
+ *
+ * LiquidBounce is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LiquidBounce is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
+ */
+package net.ccbluex.liquidbounce.features.module.modules.combat.velocity.mode
+
+import net.ccbluex.liquidbounce.config.Choice
+import net.ccbluex.liquidbounce.config.ChoiceConfigurable
+import net.ccbluex.liquidbounce.event.*
+import net.ccbluex.liquidbounce.event.events.*
+import net.ccbluex.liquidbounce.features.module.modules.combat.velocity.ModuleVelocity.modes
+import net.ccbluex.liquidbounce.features.module.modules.combat.velocity.ModuleVelocity.pause
+
+internal object VelocityDexland : Choice("Dexland") {
+
+    override val parent: ChoiceConfigurable<Choice>
+        get() = modes
+
+    val hReduce by float("HReduce", 0.3f, 0f..1f)
+    val times by int("AttacksToWork", 4, 1..10)
+
+    var lastAttackTime = 0L
+    var count = 0
+
+    val attackHandler = handler<AttackEvent> {
+        if (player.hurtTime > 0 && ++count % times == 0 && System.currentTimeMillis() - lastAttackTime <= 8000) {
+            player.velocity.x *= hReduce
+            player.velocity.z *= hReduce
+        }
+        lastAttackTime = System.currentTimeMillis()
+    }
+
+    override fun handleEvents() = super.handleEvents() && pause == 0
+
+}
